move mk_rst_table function from easybuild.tools.docs to easybuild.tools.utilities to avoid importing easybuild.tools.docs from easybuild.base.generaloption
