+++ easybuild/tools/utilities.py
@@ -40,6 +40,7 @@ from easybuild.tools.config import build_option
 
 _log = fancylogger.getLogger('tools.utilities')
 
+INDENT_4SPACES = ' ' * 4
 
 # a list of all ascii characters
 ASCII_CHARS = string.maketrans('', '')
@@ -231,3 +232,38 @@ def get_subclasses_dict(klass, include_base_class=False):
 def get_subclasses(klass, include_base_class=False):
     """Get list of all subclasses, recursively from the specified base class."""
     return get_subclasses_dict(klass, include_base_class=include_base_class).keys()
+
+
+def mk_rst_table(titles, columns):
+    """
+    Returns an rst table with given titles and columns (a nested list of string columns for each column)
+    """
+    title_cnt, col_cnt = len(titles), len(columns)
+    if title_cnt != col_cnt:
+        msg = "Number of titles/columns should be equal, found %d titles and %d columns" % (title_cnt, col_cnt)
+        raise ValueError(msg)
+    table = []
+    tmpl = []
+    line = []
+
+    # figure out column widths
+    for i, title in enumerate(titles):
+        width = max(map(len, columns[i] + [title]))
+
+        # make line template
+        tmpl.append('{%s:{c}<%s}' % (i, width))
+
+    line = [''] * col_cnt
+    line_tmpl = INDENT_4SPACES.join(tmpl)
+    table_line = line_tmpl.format(*line, c='=')
+
+    table.append(table_line)
+    table.append(line_tmpl.format(*titles, c=' '))
+    table.append(table_line)
+
+    for row in map(list, zip(*columns)):
+        table.append(line_tmpl.format(*row, c=' '))
+
+    table.extend([table_line, ''])
+
+    return table
